figureM3.svg
